+++ app/platform/[tenant]/course-content/[course_id]/course/page.tsx
@@ -1,13 +1,7 @@
 'use client';
 
-import { useEffect, useContext } from 'react';
-import { EdxIframeContext } from '@/hooks/courses/edx-iframe-context';
 import { EdxIframe } from '@/components/edx-iframe/edx-iframe';
-export default function CourseTab() {
-  const { setActiveTab } = useContext(EdxIframeContext);
-  useEffect(() => {
-    setActiveTab('course');
-  }, []);
 
+export default function CourseTab() {
   return <EdxIframe />;
 }